Refactor CompensationController and add validation
Refactor CompensationController to use constructor-based dependency injection and rename methods for clarity. Added validation for compensation description in createUpdateCompensation method and included additional fields in compensation object construction. Enhanced deleteCompensation method to check for compensation existence before deletion.

+++ src/main/java/com/example/compensationservice/Controller/CompensationController.java
@@ -3,11 +3,8 @@ package com.example.compensationservice.Controller;
 import com.example.compensationservice.Entities.Compensation;
 import com.example.compensationservice.Exceptions.CompensationException;
 import com.example.compensationservice.Service.CompensationService;
-import jakarta.validation.Valid;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -16,8 +13,11 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/compensations")
 public class CompensationController {
-    @Autowired
-    private CompensationService compensationService;
+    private final CompensationService compensationService;
+
+    public CompensationController(CompensationService compensationService) {
+        this.compensationService = compensationService;
+    }
 
     @GetMapping()
     public ResponseEntity<List<Compensation>> getAllCompensations() {
@@ -32,17 +32,21 @@ public class CompensationController {
     @PostMapping("/newCompensation")
     public ResponseEntity<String> saveCompensation(@RequestBody Compensation compensation) {
         compensationService.saveCompensation(compensation);
-        return ResponseEntity.ok( "Compensation created successfully " );
+        return ResponseEntity.ok( "Compensation created successfully" );
     }
 
-    public Compensation buildUpdateCompensation(Compensation existingCompensation, Compensation compensation) {
+    public Compensation createUpdateCompensation(Compensation existingCompensation, Compensation compensation) {
+        if(!compensation.getType().equals("salary") && (compensation.getDescription() == null || compensation.getDescription().isEmpty()))
+            throw new CompensationException("Compensation description field is required");
         return Compensation.builder()
                 .idCompensation(existingCompensation.getIdCompensation())
-                .amount(compensation.getAmount())
-                .description(compensation.getDescription())
-                .salaryAdded(compensation.getSalaryAdded())
+                .amount(compensation.getAmount() + existingCompensation.getAmount())
+                .description(compensation.getDescription() + " " + existingCompensation.getDescription())
+                .salaryAdded(existingCompensation.getSalaryAdded())
                 .type(compensation.getType())
+                .localDate(compensation.getLocalDate())
                 .dateCompensation(compensation.getDateCompensation())
+                .idUser(existingCompensation.getIdUser())
                 .build();
     }
 
@@ -51,9 +55,9 @@ public class CompensationController {
         Optional<Compensation> compensationOptional = Optional.ofNullable(compensationService.getCompensationById(id));
         if(compensationOptional.isPresent()){
             Compensation existingCompensation = compensationOptional.get();
-            Compensation updatedCompensation =  buildUpdateCompensation(existingCompensation, compensation);
+            Compensation updatedCompensation = createUpdateCompensation(existingCompensation, compensation);
             compensationService.updateCompensation(updatedCompensation);
-            return ResponseEntity.status(HttpStatus.OK).body(" User Updated Successfully");
+            return ResponseEntity.status(HttpStatus.OK).body("Compensation Updated Successfully");
         }
         throw new CompensationException("Compensation not found with id: " + id);
     }
@@ -67,6 +71,8 @@ public class CompensationController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<String> deleteCompensation(@PathVariable String id) {
+        Optional<Compensation> compensationOptional = Optional.ofNullable(compensationService.getCompensationById(id));
+        if(compensationOptional.isEmpty()) throw new CompensationException("Compensation not found compensation with id: " + id);
         compensationService.deleteCompensation(id);
         return ResponseEntity.ok("Compensation deleted successfully");
     }
